Export zosmf profile type from sdk
