+++ packages/cli/src/imperative.ts
@@ -13,9 +13,8 @@ import * as path from "path";
 
 import { IImperativeConfig } from "@zowe/imperative";
 import { Constants } from "./Constants";
-import { ZosFilesOptions } from "./zosfiles/ZosFiles.options";
 
-import { ZosmfSession } from "@zowe/zosmf-for-zowe-sdk";
+import { ZosmfProfile } from "@zowe/zosmf-for-zowe-sdk";
 import { ZosTsoProfile } from "@zowe/zos-tso-for-zowe-sdk";
 import { ZosUssProfile } from "@zowe/zos-uss-for-zowe-sdk";
 
@@ -196,108 +195,7 @@ const config: IImperativeConfig = {
         profileType: "base"
     },
     profiles: [
-        {
-            type: "zosmf",
-            schema: {
-                type: "object",
-                title: "z/OSMF Profile",
-                description: "z/OSMF Profile",
-                properties: {
-                    host: {
-                        type: "string",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_HOST
-                    },
-                    port: {
-                        type: "number",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_PORT,
-                        includeInTemplate: true
-                    },
-                    user: {
-                        type: "string",
-                        secure: true,
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_USER
-                    },
-                    password: {
-                        type: "string",
-                        secure: true,
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_PASSWORD
-                    },
-                    rejectUnauthorized: {
-                        type: "boolean",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_REJECT_UNAUTHORIZED
-                    },
-                    certFile: {
-                        type: "string",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_CERT_FILE
-                    },
-                    certKeyFile: {
-                        type: "string",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_CERT_KEY_FILE
-                    },
-                    // certFilePassphrase: {
-                    //     type: "string",
-                    //     secure: true,
-                    //     optionDefinition: ZosmfSession.ZOSMF_OPTION_CERT_FILE_PASSPHRASE
-                    // },
-                    basePath: {
-                        type: "string",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_BASE_PATH
-                    },
-                    protocol: {
-                        type: "string",
-                        optionDefinition: ZosmfSession.ZOSMF_OPTION_PROTOCOL
-                    },
-                    encoding: {
-                        type: "string",
-                        optionDefinition: {
-                            name: "encoding",
-                            aliases: ["ec"],
-                            description: "The encoding for download and upload of z/OS data set and USS files." +
-                                " The default encoding if not specified is IBM-1047.",
-                            type: "string"
-                        }
-                    },
-                    responseTimeout: {
-                        type: "number",
-                        optionDefinition: ZosFilesOptions.responseTimeout
-                    }
-                },
-                required: []
-            },
-            createProfileExamples: [
-                {
-                    options: "zos123 --host zos123 --port 1443 --user ibmuser --password myp4ss",
-                    description: "Create a zosmf profile called 'zos123' to connect to z/OSMF at host zos123 and port 1443"
-                },
-                {
-                    options: "zos124 --host zos124 --user ibmuser --password myp4ss --reject-unauthorized false",
-                    description: "Create a zosmf profile called 'zos124' to connect to z/OSMF at the host zos124 (default port - 443) " +
-                        "and allow self-signed certificates"
-                },
-                {
-                    options: "zos125 --host zos125 --port 1443",
-                    description: "Create a zosmf profile called 'zos125' to connect to z/OSMF at the host zos125 and port 1443, " +
-                        " not specifying a username or password so they are not stored on disk; these will need to be specified on every command"
-                },
-                {
-                    options: "zos126 --reject-unauthorized false",
-                    description: "Create a zosmf profile called 'zos126' to connect to z/OSMF on the default port 443 and allow self-signed " +
-                        "certificates, not specifying a username, password, or host so they are not stored on disk; these will need to be " +
-                        "specified on every command"
-                },
-                {
-                    options: "zosAPIML --host zosAPIML --port 2020 --user ibmuser --password myp4ss --reject-unauthorized false --base-path basePath",
-                    description: "Create a zosmf profile called 'zos124' to connect to z/OSMF at the host zos124 (default port - 443) " +
-                        "and allow self-signed certificates"
-                }
-            ],
-            updateProfileExamples: [
-                {
-                    options: "zos123 --user newuser --password newp4ss",
-                    description: "Update a zosmf profile named 'zos123' with a new username and password"
-                }
-            ]
-        },
+        ZosmfProfile,
         ZosTsoProfile,
         ZosUssProfile
     ]
