+++ packages/cli/src/zosfiles/ZosFiles.options.ts
@@ -10,32 +10,17 @@
 */
 
 import { ICommandOptionDefinition } from "@zowe/imperative";
+import { ZosFilesOptions } from "@zowe/zosmf-for-zowe-sdk";
 
 import i18nTypings from "./-strings-/en";
 
 // Does not use the import in anticipation of some internationalization work to be done later.
 const strings = (require("./-strings-/en").default as typeof i18nTypings).OPTIONS;
 
-/**
- * Object containing all options to be used by the Download data set API
- */
-
-export const ZosFilesOptions: { [key: string]: ICommandOptionDefinition } = {
-    /**
-     * The files timeout header value
-     * @type {ICommandOptionDefinition}
-     */
-    responseTimeout: {
-        name: "response-timeout",
-        aliases: ["rto"],
-        description: strings.RESPONSETIMEOUT,
-        type: "number",
-        defaultValue: undefined,
-        // eslint-disable-next-line @typescript-eslint/no-magic-numbers
-        numericValueRange: [5, 600]
-    }
-};
 
 export const ZosFilesOptionDefinitions: ICommandOptionDefinition[] = [
-    ZosFilesOptions.responseTimeout
+    {
+        ...ZosFilesOptions.responseTimeout,
+        description: strings.RESPONSETIMEOUT
+    }
 ];
