+++ packages/zosmf/src/index.ts
@@ -11,6 +11,7 @@
 
 export * from "./constants/Zosmf.constants";
 export * from "./constants/Zosmf.messages";
+export * from "./constants/Zosmf.profile";
 
 export * from "./doc/IDefinedSystem";
 export * from "./doc/IZosmfInfoResponse";
